feature(feature/ui-product-create-files): se modifica vista de creacion de producto para cargar archivos
